:feature  agregar typescript al proyecto
