+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "inventory-app",
-  "version": "0.0.0",
+  "version": "1.0.0",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
@@ -20,7 +20,10 @@
     "wouter": "^2.8.0-alpha.2"
   },
   "devDependencies": {
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
     "@vitejs/plugin-react": "^1.0.7",
+    "typescript": "^4.9.5",
     "vite": "^2.7.2"
   }
 }
