--- src/components/inventory/Deletedialog.jsx
@@ -1,46 +1,0 @@
-import { useRef} from "react";
-import {
-  AlertDialog,
-  AlertDialogBody,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogContent,
-  AlertDialogOverlay,
-  Button
-} from '@chakra-ui/react'
-
-export function Deletedialog({sku,isOpen,onClose,handleDelete}) {
-    const cancelRef = useRef();
-    const handel=()=>handleDelete(sku);
-    
-    return (
-        <>
-            <AlertDialog
-            isOpen={isOpen}
-            leastDestructiveRef={cancelRef}
-            onClose={onClose}
-          >
-            <AlertDialogOverlay>
-              <AlertDialogContent>
-                <AlertDialogHeader fontSize='lg' fontWeight='bold'>
-                  Delete Product
-                </AlertDialogHeader>
-    
-                <AlertDialogBody>
-                  Are you sure delete? You can't undo this action afterwards.
-                </AlertDialogBody>
-    
-                <AlertDialogFooter>
-                  <Button ref={cancelRef} onClick={onClose}>
-                    Cancel
-                  </Button>
-                  <Button colorScheme='red' onClick={handel} ml={3}>
-                    Delete
-                  </Button>
-                </AlertDialogFooter>
-              </AlertDialogContent>
-            </AlertDialogOverlay>
-          </AlertDialog>
-        </>
-      )
-}
